@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
 
@@ -28,7 +28,7 @@
 
   <ItemGroup>
     <Reference Include="NetworkMessage">
-      <HintPath>..\..\NetworkMessage\NetworkMessage\bin\Debug\net6.0\NetworkMessage.dll</HintPath>
+      <HintPath>..\..\NetworkMessage\NetworkMessage\bin\Debug\net8.0\NetworkMessage.dll</HintPath>
     </Reference>
   </ItemGroup>
 
