@@ -2,16 +2,52 @@
 @{
     Layout = "_AccountLayout";
 }
+
 <div class="container">
-    <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-    <form asp-action="Authorization" asp-controller="Account" method="post" >
-        <input type="email" asp-for="Email" class="form-control mb-3" placeholder="Введите свою почту" maxlength="20" />
-        <span asp-validation-for="Email" class="text-danger"></span>
-        <input type="password" asp-for="Password" class="form-control mb-3" placeholder="Введите свой пароль" maxlength="20" />
-        <span asp-validation-for="Password" class="text-danger"></span>
-        <input type="submit" value="Войти" class="btn btn-primary" />
-    </form>
-    <a onclick='window.location.href = "@Url.Action("Registration", "Account")"' class="btn btn-link">Зарегистрироваться</a>
+	<div class="row">
+		<div class="col-md-8 offset-md-2">
+			<div class="wrapper d-flex align-items-center justify-content-center" style="height: 100vh">
+				<div class="card flex-fill" style="height: 70vh">
+					<div class="card-body">
+						<div class="row d-flex align-items-center justify-content-center h-100">
+							<div class="col-md-5 offset-md-1">
+								<h2 class="custom-header text-center">Войти</h2>
+								<div asp-validation-summary="ModelOnly" class="text-danger"></div>
+								<form asp-controller="Account" asp-action="Authorization" method="post" asp-route-returnUrl="@Model.ReturnUrl">
+									<div class="mb-3">
+										<label for="email" class="form-label">Почта</label>
+										<input type="text" name="email" id="email" class="form-control" asp-for="Email" placeholder="example.email@gmail.com">
+										<span asp-validation-for="Email" class="text-danger"></span>
+									</div>
+
+									<div class="mb-3">
+										<label for="password" class="form-label">Пароль</label>
+										<input type="text" name="password" id="password" class="form-control" asp-for="Password">
+										<span asp-validation-for="Password" class="text-danger"></span>
+									</div>
+
+									<div class="d-flex justify-content-center">
+										<div class="btn-group">
+											<input type="submit" class="btn btn-primary" value="Войти">
+											<a onclick='window.location.href = "@Url.Action("Registration", "Account")"' class="btn btn-outline-secondary">Регистрация</a>
+										</div>
+									</div>
+								</form>
+
+								<div class="text-center mt-4">
+									<span>Войти с помощью</span><br>
+									<button class="btn"><i class="fa fa-google"></i></button>
+								</div>
+							</div>
+							<div class="col-md-5 d-md-block d-none">
+								<img src="/img/authImg.png" alt="" class="img-fluid">
+							</div>
+						</div>
+					</div>
+				</div>
+			</div>
+		</div>
+	</div>
 </div>
 
 
