@@ -1,3 +1,3 @@
 ﻿@using RemoteControlServer
-@using RemoteControlServer.Models
+@using RemoteControlServer.ViewModels
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
